refactor: move resolver and mutator types in separated file
