--- src/infra/communication/observer.ts
@@ -1,5 +1,0 @@
-export interface Observer<T> {
-  onAdd(value: T): void
-  onUpdate(value: T): void
-  onRemove(value: T): void
-}
